feat: pre-release

+++ WeatherForecast.Domain/Aggregates/WeatherForecast/Commands/IAddWeatherForecastCommand.cs
@@ -8,11 +8,11 @@ public interface IAddWeatherForecastCommand
     public DateTime DateWeatherEvent { get; }
     public string CityName { get; }
     public double Temperature { get; }
-    public double HumidityInPercent { get; }
+    public int HumidityInPercent { get; }
     public double DewPoint { get; }
     public int AtmospherePressure { get; }
     public IAddWindValueCommand Wind { get; }
-    public double CloudinessInPercent { get; }
+    public int CloudinessInPercent { get; }
     public int CloudBaseInMeters { get; }
     public int HorizontalVisibilityInKilometer { get; }
     public string WeatherEvent { get; }
